% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_elevations.R
\docType{data}
\name{data_url}
\alias{data_url}
\title{Function to download D-PLACE elevations data flat files and load it as a dataframe}
\format{
An object of class \code{character} of length 1.
}
\usage{
data_url
}
\value{
a dataframe object
}
\description{
Function to download D-PLACE elevations data flat files and load it as a dataframe
}
\keyword{datasets}
