% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplace_societies.R
\name{dplace_all_societies}
\alias{dplace_all_societies}
\title{Get data about all societies listed in D-Place}
\usage{
dplace_all_societies(path = "societies")
}
\arguments{
\item{path}{character, address in the API to call}
}
\value{
data.frame
}
\description{
Get data about all societies listed in D-Place (d-place.org)
}
