% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplace_societies.R
\name{dplace_all_societies_}
\alias{dplace_all_societies_}
\title{Get one page of results about societies from D-Place}
\usage{
dplace_all_societies_(path = NULL, query = NULL)
}
\arguments{
\item{path}{character, address in the API to call}

\item{query}{character, which page to return. Format: "page=X"}
}
\value{
data.frame
}
\description{
Use \code{\link{dplace_all_societies}} instead.
}
