% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dplace_getter}
\alias{dplace_getter}
\title{Function to download and process a page of data from D-Place
#' @param path character, address in the API to call}
\usage{
dplace_getter(path = NULL, query = NULL)
}
\arguments{
\item{query}{character, which page to return. Format: "page=X"}
}
\value{
list of processed JSON
}
\description{
Function to download and process a page of data from D-Place
#' @param path character, address in the API to call
}
