% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplace_society.r
\name{dplace_society}
\alias{dplace_society}
\title{Get data about a society}
\usage{
dplace_society(ext_id = NULL, path = "society")
}
\arguments{
\item{ext_id}{character, ID of the society}

\item{path}{character, address in the API to call}
}
\value{
list containing society data, including `env` and `trait` objects with variables and values
}
\description{
Get data about a particular society listed in D-Place (d-place.org)
}
