% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplacer.r
\docType{package}
\name{dplacer}
\alias{dplacer}
\alias{dplacer-package}
\title{dplacer - an R interface to the anthropology database, [D-Place](https://d-place.org).}
\description{
dplacer - an R interface to the anthropology database, [D-Place](https://d-place.org).
}
