% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_geography.R
\docType{data}
\name{elevation_data_url}
\alias{elevation_data_url}
\title{Function to download D-PLACE geography data flat files and load it as a dataframe}
\format{
An object of class \code{character} of length 1.
}
\usage{
elevation_data_url
}
\value{
a dataframe object
}
\description{
Function to download D-PLACE geography data flat files and load it as a dataframe
}
\keyword{datasets}
