% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_trees.R
\name{load_trees}
\alias{load_trees}
\title{Loads trees from DPLACE}
\usage{
load_trees(dirname, type = "summary")
}
\arguments{
\item{type}{the type of tree to load (\code{summary} or \code{posterior})}
}
\value{
An ape::multiPhylo object
}
\description{
Loads trees from DPLACE
}
\examples{
trees <- load_trees("kolipakam_et_al2018")
}
