% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_Binford.R, R/download_EA.R
\docType{data}
\name{variables_url}
\alias{variables_url}
\title{Function to download Binford D-PLACE data flat files and load it as a dataframe}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
variables_url

variables_url
}
\value{
a dataframe object

a dataframe object
}
\description{
Function to download Binford D-PLACE data flat files and load it as a dataframe

Function to download EA D-PLACE data flat files and load it as a dataframe
}
\keyword{datasets}
