% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{dp_search_variables}
\alias{dp_search_variables}
\title{Search D-PLACE variables by keyword}
\usage{
dp_search_variables(keyword)
}
\arguments{
\item{keyword}{A search term matched (case-insensitively) against
variable names and descriptions.}
}
\value{
A tibble of matching variables (see [dplace_variables] for column
  definitions).
}
\description{
A thin, more discoverable wrapper around `dp_variables(search = keyword)`.
}
\examples{
dp_search_variables("marriage")

}
