% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_values.R
\name{dp_values}
\alias{dp_values}
\title{Look up raw coded values}
\usage{
dp_values(var_id = NULL, soc_id = NULL)
}
\arguments{
\item{var_id}{Optional character vector of variable ID(s) to filter to.}

\item{soc_id}{Optional character vector of society ID(s) to filter to.}
}
\value{
A tibble of values (see [dplace_values] for column definitions).
}
\description{
Filter the long-format D-PLACE data table ([dplace_values]) by variable
and/or society. For an analysis-ready table that also joins in society
information and (for categorical/ordinal variables) code labels, use
[dp_variable_data()] instead.
}
\examples{
dp_values(var_id = "EA202", soc_id = "Sa1")

}
