% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplace_variables.r
\name{dplace_variables}
\alias{dplace_variables}
\title{Get list of cultural variables}
\usage{
dplace_variables(path = "variables")
}
\description{
To do:
[ ] the 'index_categories' column is currently a jagged list; convert it to separate columns
}
